[ALPS09686298] Coverity fix for gpu dvfs kernel 6.12 code
[Description]
Coverity fix for gpu dvfs kernel 6.12 code
Rework for Negative loop bound
-> Type change for the variable of loop exit condition

[Test]
Local build pass & boot up pass

MTK-Commit-Id: 514a00de05fc792845b92b9f54c56c4d6298174d

CR-Id: ALPS09686298
Feature: GPU DVFS Policy

+++ drivers/gpu/mediatek/ged/src/ged_dvfs.c
@@ -4079,7 +4079,7 @@ int ged_dvfs_query_opp_cost(struct GED_DVFS_OPP_STAT *psReport,
 		gpu_opp_logs_enable == 1) {
 
 		if (is_fdvfs_enable() & POLICY_MODE_V2) {
-			for (unsigned int i = 0; i < real_opp_num; i++) {
+			for (int i = 0; i < (int)real_opp_num; i++) {
 				if (i >= i32NumOpp)
 					report_idx = i32NumOpp - 1;
 				else
